Dchassin/docker warnings (#182)
* Fixed warnings in climate
* Fixed warnings in commercial module
* Removed unsupported items as 4.2 from configure.ac
* Fixed warnings in generator module
* Fixed warnings in market module
* Fixed error in climate module
* Fixed warnings in mysql
* Fixed powerflow module warnings
* Fixed warnings in residential module
* Fixed error in climate/csv_reader.cpp
* Fixed warnings in tape module
* Fixed warnings in gldcore

+++ powerflow/regulator.cpp
@@ -1366,7 +1366,7 @@ int regulator::kmldata(int (*stream)(const char*,...))
 
 	// tap position
 	stream("<TR><TH ALIGN=LEFT>Tap position</TH>");
-	for ( int i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
+	for ( size_t i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
 	{
 		if ( phase[i] )
 			stream("<TD ALIGN=CENTER COLSPAN=2 STYLE=\"font-family:courier;\"><NOBR>%d</NOBR></TD>", tap[i]);
@@ -1382,7 +1382,7 @@ int regulator::kmldata(int (*stream)(const char*,...))
 	if ( run_realtime.get_bool() )
 	{
 		stream("<TR><TH ALIGN=LEFT>Raise to</TH>");
-		for ( int i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
+		for ( size_t i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
 		{
 			if ( phase[i] )
 				stream("<TD ALIGN=CENTER COLSPAN=2 STYLE=\"font-family:courier;\"><FORM ACTION=\"http://%s:%d/kml/%s\" METHOD=GET><INPUT TYPE=SUBMIT NAME=\"tap_%c\" VALUE=\"%d\" /></FORM></TD>",
@@ -1392,7 +1392,7 @@ int regulator::kmldata(int (*stream)(const char*,...))
 		}
 		stream("</TR>\n");
 		stream("<TR><TH ALIGN=LEFT>Lower to</TH>");
-		for ( int i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
+		for ( size_t i = 0 ; i<sizeof(phase)/sizeof(phase[0]) ; i++ )
 		{
 			if ( phase[i] )
 				stream("<TD ALIGN=CENTER COLSPAN=2 STYLE=\"font-family:courier;\"><FORM ACTION=\"http://%s:%d/kml/%s\" METHOD=GET><INPUT TYPE=SUBMIT NAME=\"tap_%c\" VALUE=\"%d\" /></FORM></TD>",
